images, main info, sizes and colors refactored
