+++ assets/mock-data/product_detail.json
@@ -7,6 +7,8 @@
   "before": 299.43,
   "offer": 299.43,
   "rating" : 3.5,
+  "sizes" : ["6", "6.5", "7", "7.5", "8", "8.5", "9"],
+  "colors" : ["0xFFFFC833", "0xFF40BFFF", "0xFFFB7181", "0xFF53D1B6", "0xFF5C61F4", "0xFF223263"],
   "description" : "The Nike Air Max 270 React ENG combines a full-length React foam midsole with a 270 Max Air unit for unrivaled comfort and a striking visual experience.",
   "related_products" : [
     {
