+++ lib/src/modules/product-details/models/product_detail_model.dart
@@ -1,40 +1,46 @@
 class ProductDetailModel {
   final String id;
   final String title;
-  final String image1;
-  final String image3;
-  final String image2;
+  final String firstImage;
+  final String thirdImage;
+  final String secondImage;
   final double price;
   final double before;
   final double offer;
   final String description;
   final double rating;
+  final List<String> sizes;
+  final List<String> colors;
 
   const ProductDetailModel({
     required this.title,
     required this.price,
-    required this.image1,
+    required this.firstImage,
     required this.id,
-    required this.image2,
-    required this.image3,
+    required this.secondImage,
+    required this.thirdImage,
     required this.before,
     required this.offer,
     required this.description,
     required this.rating,
+    required this.sizes,
+    required this.colors,
   });
 
   factory ProductDetailModel.fromJson(Map<String, dynamic> json) {
     return ProductDetailModel(
       id: json['id'] as String,
       title: json['title'] as String,
-      image1: json['image1'] as String,
-      image2: json['image2'] as String,
-      image3: json['image3'] as String,
+      firstImage: json['image1'] as String,
+      secondImage: json['image2'] as String,
+      thirdImage: json['image3'] as String,
       price: json['price'] as double,
       before: json['before'] as double,
       offer: json['offer'] as double,
       description: json['description'] as String,
       rating: json['rating'] as double,
+      sizes: (json['sizes'] as List).map((e) => e as String).toList(),
+      colors: (json['colors'] as List).map((e) => e as String).toList(),
     );
   }
 }
